+++ packages/json/src/components/index.ts
@@ -1,5 +1,5 @@
+export * from "./json-array.jsx";
+export * from "./json-object.jsx";
 export * from "./json-value.jsx";
-export * from "./JsonArray.jsx";
-export * from "./JsonObject.jsx";
-export * from "./Reference.jsx";
+export * from "./reference2.jsx";
 export * from "./SourceFile.jsx";
